Make DuplicateSelection into a Scene extension method
